requirements.txt
# requirements.txt (for CPU Fallback Plan)

# Pin PyTorch/Torchaudio/NumPy
torch==2.3.1
torchaudio==2.3.1
numpy<2.0

# Core ASR and Diarization - Install from PyPI
faster-whisper>=1.0.0 # Let pip handle dependencies including ctranslate2
pyannote.audio>=3.1.1,<3.4 # Keep pinned pyannote

# Other dependencies
speechbrain>=1.0.0
faiss-cpu>=1.7.4
transformers>=4.48.0
optree>=0.13.0

# Utilities
soundfile>=0.12.1
librosa>=0.10.0
pandas>=1.5.0
tqdm>=4.62.3
requests>=2.28.0
huggingface-hub>=0.17.0
ffmpeg-python>=0.2.0
av>=10.0.0 # Still needed by faster-whisper


